geo_utils splitting
