+++ pygeo/geo_utils/__init__.py
@@ -1,0 +1,31 @@
+# =============================================================================
+# Utility Functions for Use in pyNetwork, pyGeo, pyBlock, DVGeometry,
+# and pyLayout
+# =============================================================================
+
+# This __init__ file imports every methods in pygeo/geo_utils
+from .bilinear_map import *  # noqa: F401, F403
+from .dcel import *  # noqa: F401, F403
+from .file_io import *  # noqa: F401, F403
+from .index_position import *  # noqa: F401, F403
+from .knowvector import *  # noqa: F401, F403
+from .misc import *  # noqa: F401, F403
+from .node_edge_face import *  # noqa: F401, F403
+from .norm import *  # noqa: F401, F403
+from .orientation import *  # noqa: F401, F403
+from .pointselect import *  # noqa: F401, F403
+from .polygon import *  # noqa: F401, F403
+from .projection import *  # noqa: F401, F403
+from .remove_duplicates import *  # noqa: F401, F403
+from .rotation import *  # noqa: F401, F403
+from .split_quad import *  # noqa: F401, F403
+from .topology import *  # noqa: F401, F403
+from .tripan import *  # noqa: F401, F403
+
+
+# Set a (MUCH) larger recursion limit. For meshes with extremely large
+# numbers of blocs (> 5000) the recursive edge propagation may hit a
+# recursion limit.
+import sys
+
+sys.setrecursionlimit(10000)
